pyproject.toml
[project]
name = "invoice-engine"
version = "0.1.0"
description = "A robust ETL pipeline for legacy invoice data archaeology."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "olefile>=0.47",
    "python-dotenv>=1.2.1",
]

[project.scripts]
explore-doc-inspector = "v2.exploration.doc_inspector:main"
run-discovery = "v2.discovery.engine:run_discovery"
run-conversion = "v2.conversion.engine:run_conversion"
run-pipeline = "v2.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["v2"]
